Fixed compiler's conversion warnings

+++ Player.cpp
@@ -32,22 +32,22 @@ void Player::handleInput()
 	{
 		if ((TheInputHandler::Instance()->xvalue(0, 1) > 0) || (TheInputHandler::Instance()->xvalue(0, 1) < 0))
 		{
-			m_velocity.setX(1 * TheInputHandler::Instance()->xvalue(0, 1));
+			m_velocity.setX((float)1 * TheInputHandler::Instance()->xvalue(0, 1));
 		}
 
 		if ((TheInputHandler::Instance()->yvalue(0, 1) > 0) || (TheInputHandler::Instance()->yvalue(0, 1) < 0))
 		{
-			m_velocity.setY(1 * TheInputHandler::Instance()->yvalue(0, 1));
+			m_velocity.setY((float)1 * TheInputHandler::Instance()->yvalue(0, 1));
 		}
 
 		if ((TheInputHandler::Instance()->xvalue(0, 2) > 0) || (TheInputHandler::Instance()->xvalue(0, 2) < 0))
 		{
-			m_velocity.setX(1 * TheInputHandler::Instance()->xvalue(0, 2));
+			m_velocity.setX((float)1 * TheInputHandler::Instance()->xvalue(0, 2));
 		}
 
 		if ((TheInputHandler::Instance()->yvalue(0, 2) > 0) || (TheInputHandler::Instance()->yvalue(0, 2) < 0))
 		{
-			m_velocity.setY(1 * TheInputHandler::Instance()->yvalue(0, 2));
+			m_velocity.setY((float)1 * TheInputHandler::Instance()->yvalue(0, 2));
 		}
 	}
 }
